requirements.txt
Authlib==1.3.0
blinker==1.7.0
cachelib==0.13.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
Flask==3.0.3
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
oauthlib==2.1.0
pycparser==2.22
pysqlite3==0.5.2
python-dotenv==1.0.1
requests==2.31.0
requests-oauthlib==1.1.0
urllib3==2.2.1
Werkzeug==3.0.2
